Rename SkuSelector files
